requirements.txt
Pillow==8.4.0
pydicom==2.3.0
matplotlib==3.4.3